+++ site/owid.entry.ts
@@ -9,7 +9,7 @@ import { runChartsIndexPage } from "./runChartsIndexPage.js"
 import { runSearchPage } from "./search/SearchPanel.js"
 import { runNotFoundPage } from "./NotFoundPageMain.js"
 import { runFeedbackPage } from "./Feedback.js"
-import { runDonateForm } from "./stripe/DonateForm.js"
+import { runDonateForm } from "./DonateForm.js"
 import { runCountryProfilePage } from "./runCountryProfilePage.js"
 import { runTableOfContents } from "./TableOfContents.js"
 import { runRelatedCharts } from "./blocks/RelatedCharts.js"
