--- site/stripe/readme.md
@@ -1,3 +1,0 @@
-# Stripe
-
-This folder contains the donation form and code for interacting with the Stripe API.
